\name{tensorApp-package}
\alias{tensorApp-package}
\alias{tensorApp}
\docType{package}
\title{
  High-order SVD approximation by Tucker and CP decomposition and selection of ranks
}
\description{
  High-order SVD approximation by Tucker decomposition or CANDECOMP/PARAFAC (CP) decomposition and selection of ranks. 
}
\details{
  High-order SVD approximation by Tucker decomposition or CANDECOMP/PARAFAC (CP) decomposition and selection of ranks.
}
\author{
Xu Liu

Maintainer: Xu Liu <liu.xu@sufe.edu.cn>
}

\keyword{CP decomposition; HOSVD; Tucker decomposition}
